+++ lib/nodeElement/ghostDropHint.js
@@ -13,14 +13,14 @@ function _defineProperty(obj, key, value) { key = _toPropertyKey(key); if (key i
 function _toPropertyKey(arg) { var key = _toPrimitive(arg, "string"); return _typeof(key) === "symbol" ? key : String(key); }
 function _toPrimitive(input, hint) { if (_typeof(input) !== "object" || input === null) return input; var prim = input[Symbol.toPrimitive]; if (prim !== undefined) { var res = prim.call(input, hint || "default"); if (_typeof(res) !== "object") return res; throw new TypeError("@@toPrimitive must return a primitive value."); } return (hint === "string" ? String : Number)(input); }
 var GhostDropHint = /*#__PURE__*/function () {
-  function GhostDropHint(node, $element, position) {
+  function GhostDropHint(node, element, position) {
     _classCallCheck(this, GhostDropHint);
-    _defineProperty(this, "$element", void 0);
+    _defineProperty(this, "element", void 0);
     _defineProperty(this, "node", void 0);
-    _defineProperty(this, "$ghost", void 0);
-    this.$element = $element;
+    _defineProperty(this, "ghost", void 0);
+    this.element = element;
     this.node = node;
-    this.$ghost = jQuery("<li class=\"jqtree_common jqtree-ghost\"><span class=\"jqtree_common jqtree-circle\"></span>\n          <span class=\"jqtree_common jqtree-line\"></span></li>");
+    this.ghost = this.createGhostElement();
     if (position === _node.Position.After) {
       this.moveAfter();
     } else if (position === _node.Position.Before) {
@@ -36,17 +36,17 @@ var GhostDropHint = /*#__PURE__*/function () {
   _createClass(GhostDropHint, [{
     key: "remove",
     value: function remove() {
-      this.$ghost.remove();
+      this.ghost.remove();
     }
   }, {
     key: "moveAfter",
     value: function moveAfter() {
-      this.$element.after(this.$ghost);
+      this.element.after(this.ghost);
     }
   }, {
     key: "moveBefore",
     value: function moveBefore() {
-      this.$element.before(this.$ghost);
+      this.element.before(this.ghost);
     }
   }, {
     key: "moveInsideOpenFolder",
@@ -54,14 +54,27 @@ var GhostDropHint = /*#__PURE__*/function () {
       var _this$node$children$;
       var childElement = (_this$node$children$ = this.node.children[0]) === null || _this$node$children$ === void 0 ? void 0 : _this$node$children$.element;
       if (childElement) {
-        jQuery(childElement).before(this.$ghost);
+        childElement.before(this.ghost);
       }
     }
   }, {
     key: "moveInside",
     value: function moveInside() {
-      this.$element.after(this.$ghost);
-      this.$ghost.addClass("jqtree-inside");
+      this.element.after(this.ghost);
+      this.ghost.classList.add("jqtree-inside");
+    }
+  }, {
+    key: "createGhostElement",
+    value: function createGhostElement() {
+      var ghost = document.createElement("li");
+      ghost.className = "jqtree_common jqtree-ghost";
+      var circleSpan = document.createElement("span");
+      circleSpan.className = "jqtree_common jqtree-circle";
+      ghost.append(circleSpan);
+      var lineSpan = document.createElement("span");
+      lineSpan.className = "jqtree_common jqtree-line";
+      ghost.append(lineSpan);
+      return ghost;
     }
   }]);
   return GhostDropHint;
