+++ lib/nodeElement/index.js
@@ -44,7 +44,7 @@ var NodeElement = /*#__PURE__*/function () {
       if (this.mustShowBorderDropHint(position)) {
         return new _borderDropHint["default"](this.element, this.treeWidget._getScrollLeft());
       } else {
-        return new _ghostDropHint["default"](this.node, jQuery(this.element), position);
+        return new _ghostDropHint["default"](this.node, this.element, position);
       }
     }
   }, {
